feat: implement SemestresFrame for semester management interface
